requirements.txt
python-dotenv
watchdog
PyQt6 
mysql-connector-python